requirements.txt
Flask~=0.12.2
Flask-SQLAlchemy~=2.2
Flask-Migrate~=2.1.1
gunicorn~=19.7.1
geoalchemy2~=0.4.0
psycopg2~=2.7.3
python-dotenv~=0.7.1
